requirements.txt
opencv-contrib-python
lxml
tqdm
tensorflow==2.3.0rc0
absl-py
easydict~=1.9
matplotlib~=3.3.2
pillow~=8.0.1
playsound~=1.2.2

numpy~=1.18.5
scipy~=1.4.1